+++ include/editor.h
@@ -83,6 +83,7 @@ public:
     void addNewConnection();
     void removeCurrentConnection();
     void addNewWildMonGroup(QWidget *window);
+    void deleteWildMonGroup();
     void updateDiveMap(QString mapName);
     void updateEmergeMap(QString mapName);
     void setSelectedConnectionFromMap(QString mapName);
@@ -211,7 +212,6 @@ private slots:
     void onHoveredMapMovementPermissionCleared();
     void onSelectedMetatilesChanged();
     void onWheelZoom(int);
-    void onMapRulerLengthChanged();
 
 signals:
     void objectsChanged();
@@ -220,6 +220,7 @@ signals:
     void wildMonDataChanged();
     void warpEventDoubleClicked(QString mapName, QString warpNum);
     void currentMetatilesSelectionChanged();
+    void mapRulerStatusChanged(const QString &);
 };
 
 #endif // EDITOR_H